Add comment level
